requirements.txt
pytest==7.2.2
requests==2.31.0
allure-pytest==2.13.2
beautifulsoup4==4.10.0
pyyaml==6.0